// functions/src/types/video.ts
export interface VideoFromKS {
  mediaType: number
  conversionQuality: number
  sourceType: string
  dataUrl: string
  flavorParamsIds: string
  plays: number
  views: number
  lastPlayedAt: number
  width: number
  height: number
  duration: number
  msDuration: number
  id: string
  name: string
  description: string
  partnerId: number
  userId: string
  creatorId: string
  tags: string
  categories: string
  categoriesIds: string
  status: number
  moderationStatus: number
  moderationCount: number
  type: number
  createdAt: number
  updatedAt: number
  rank: number
  totalRank: number
  votes: number
  downloadUrl: string
  searchText: string
  licenseType: number
  version: number
  thumbnailUrl: string
  accessControlId: number
  replacementStatus: number
  partnerSortValue: number
  conversionProfileId: number
  rootEntryId: string
  operationAttributes: any[]
  entitledUsersEdit: string
  entitledUsersPublish: string
  entitledUsersView: string
  capabilities: string
  displayInSearch: number
  application: string
  applicationVersion: string
  blockAutoTranscript: boolean
  objectType: string
}